Starting to move uvfits tests to separate file
